requirements.txt
python-can==4.4.2
pyserial==3.5